Fix a bunch of modern gcc warnings, and remove obsolete basic crypto.
The basic crypto encoders were producing openssl deprecation warnings
and are extremely dangerous to use anyway (never roll your own
crypto!), so haven't actually been used in any actual programs for a
long time.

+++ streams/t/wvstream.t.cc
@@ -59,7 +59,7 @@ public:
     
     virtual size_t uwrite(const void *buf, size_t count)
     {
-	fprintf(stderr, "I'm uwrite! (%d)\n", count);
+	fprintf(stderr, "I'm uwrite! (%d)\n", (int)count);
 	if (block_writes)
 	    return 0;
 	size_t ret = WvStream::uwrite(buf, count);
